backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
aiohttp==3.9.5
aiodns==3.0.0
dnspython==2.6.1
python-whois==0.8.0
python-dateutil==2.9.0.post0
orjson==3.10.3
pydantic-settings==2.2.1
pytest==8.1.1
pytest-asyncio==0.23.6
idna==3.6
asyncpg==0.29.0
